+++ NEWS.md
@@ -4,7 +4,7 @@ Bugfixes:
 
 * Plotting functions no longer fail with incomplete source data.
 * Plotting functions now respect the order of `modx.values` and `mod2.values`
-arguments.
+arguments. (#29)
 
 # interactions 1.1.1